requirements.txt
aiogram==2.25.1
python-dotenv==1.0.1
pytest==8.2.0
python-docx==1.1.0
fpdf2==2.7.9